+++ Console Edition/src/Installer/Installator.vb
@@ -16,7 +16,7 @@
 
 
 
-    Sub Installator(ByRef checkpath As Object(), ByRef nservers As Integer, ByRef nameservers As String(), ByRef installationstatus As String(), ByRef versionstatus As String(), ByRef path As String(), ByRef numberservers As String(), _
+    Sub Installator(ByRef checkpath As Object(), ByRef nservers As SByte, ByRef nameservers As String(), ByRef installationstatus As String(), ByRef versionstatus As String(), ByRef path As String(), ByRef numberservers As String(), _
                     ByRef back As String, ByRef menu1 As String, ByRef changemade As String, ByRef status1 As String, ByRef version1 As String, ByRef versionstable1 As String, ByRef versionbeta1 As String, ByRef versiondev1 As String, _
                     ByRef versionsoft1 As String, ByRef currentversion As String, ByRef writepath1 As String, ByRef writepath2 As String, ByRef writepath3 As String, ByRef installatortitle As String, ByRef installator1 As String, ByRef installator2 As String, _
                     ByRef installator3 As String, ByRef installator4 As String, ByRef installator5 As String, ByRef installator6 As String, ByRef installator7 As String, ByRef versions1 As String, ByRef versions2 As String, ByRef versions3 As String)
@@ -36,23 +36,23 @@
 
         Do
             Do
-                    Console.Clear()
-                    Console.ForegroundColor = ConsoleColor.Green
-                    Console.WriteLine("========================<PocketMine Manager Servers>============================")
-                    Console.ForegroundColor = ConsoleColor.Cyan
-                    Console.WriteLine("{0}", installatortitle)
-                    Console.ForegroundColor = ConsoleColor.White
-                    For i = 1 To nservers
-                        Console.WriteLine("{0}) {1} -> {2}: {3} -> {4}: {5}", i, nameservers(i - 1), version1, versionstatus(i - 1), status1, installationstatus(i - 1))
+                Console.Clear()
+                Console.ForegroundColor = ConsoleColor.Green
+                Console.WriteLine("========================<PocketMine Manager Servers>============================")
+                Console.ForegroundColor = ConsoleColor.Cyan
+                Console.WriteLine("{0}", installatortitle)
+                Console.ForegroundColor = ConsoleColor.White
+                For i = 1 To nservers
+                    Console.WriteLine("{0}) {1} -> {2}: {3} -> {4}: {5}", i, nameservers(i - 1), version1, versionstatus(i - 1), status1, installationstatus(i - 1))
 
-                    Next
+                Next
                 Console.WriteLine("11) {0}", back)
-                    Console.WriteLine()
+                Console.WriteLine()
                 Console.Write("{0} ", installator2)
-                    chooseserver = Console.ReadLine
+                chooseserver = Console.ReadLine
 
-                    If chooseserver = "11" Then
-                        Exit Sub
+                If chooseserver = "11" Then
+                    Exit Sub
                 End If
             Loop While chooseserver < "1" Or chooseserver > "10"
