+++ Console Edition/src/Installer/ManagerInstaller.vb
@@ -13,7 +13,7 @@
     'it under the terms of the GNU Lesser General Public License as published by 
     'the Free Software Foundation, either version 3 of the License, or 
     '(at your option) any later version. 
-    Sub ManagerInstaller(ByRef path As String(), ByRef nameservers As String(), ByRef nservers As Integer, ByRef checkpath As Object(), ByRef numberservers As String(), ByRef downloadstatus As String(), ByRef installationstatus As String(), ByRef versionstatus As String(), ByRef checknameserver As Object(), _
+    Sub ManagerInstaller(ByRef path As String(), ByRef nameservers As String(), ByRef nservers As SByte, ByRef checkpath As Object(), ByRef numberservers As String(), ByRef downloadstatus As String(), ByRef installationstatus As String(), ByRef versionstatus As String(), ByRef checknameserver As Object(), _
                          ByRef back As String, ByRef changemade As String, ByRef status1 As String, ByRef version1 As String, ByRef versionstable1 As String, ByRef versionbeta1 As String, ByRef versiondev1 As String, ByRef versionsoft1 As String, ByRef currentversion As String, ByRef writepath1 As String, _
                          ByRef writepath2 As String, ByRef writepath3 As String, ByRef menu1 As String, ByRef installertitle As String, ByRef installer1 As String, ByRef installer2 As String, ByRef installer3 As String, ByRef installatortitle As String, ByRef installator1 As String, ByRef installator2 As String, ByRef installator3 As String, ByRef installator4 As String, ByRef installator5 As String, _
                          ByRef installator6 As String, ByRef installator7 As String, ByRef downloadertitle As String, ByRef downloader1 As String, ByRef downloader2 As String, ByRef downloader3 As String, ByRef downloader4 As String, ByRef downloader5 As String, ByRef downloader6 As String, ByRef downloader7 As String, ByRef downloader8 As String, _
@@ -65,8 +65,8 @@
 
     Sub ChangeVersionStatus(ByRef version As String, ByRef chooseserver As String, ByRef versionstatus As String(), ByRef changemade As String)
 
-        Dim index As Integer
-        Dim indexstatus As Integer = -1
+        Dim index As SByte
+        Dim indexstatus As SByte = -1
         Dim status As String() = New String() {"Stable", "Beta", "Dev", "Soft"}
         Dim verified As Boolean = False
